+++ src/app/static/src/app/hintVersion.ts
@@ -1,2 +1,1 @@
-export const currentHintVersion = "1.64.3";
-
+export const currentHintVersion = "1.65.0";