+++ src/app/static/src/tests/root/actions.test.ts
@@ -3,12 +3,15 @@ import {
     mockAxios,
     mockFailure,
     mockBaselineState,
+    mockError,
+    mockGenericChartState,
     mockHintrVersionState,
     mockModelCalibrateState,
     mockModelRunState,
     mockProjectsState,
     mockRootState,
-    mockStepperState, mockSuccess, mockError
+    mockStepperState,
+    mockSuccess
 } from "../mocks";
 import {Language} from "../../app/store/translations/locales";
 import {LanguageMutation} from "../../app/store/language/mutations";
@@ -453,6 +456,23 @@ describe("root actions", () => {
         expect(dispatch.mock.calls.length).toBe(1);
         expect(dispatch.mock.calls[0][0]).toStrictEqual("metadata/getPlottingMetadata");
         expect(dispatch.mock.calls[0][1]).toStrictEqual("MWI");
+    });
+
+    it("changeLanguage refreshes genericChart datasets, if any", async() => {
+        const commit = jest.fn();
+        const dispatch = jest.fn();
+        const rootState = mockRootState(
+            {
+                baseline: mockBaselineState({iso3: "MWI"}),
+                genericChart: mockGenericChartState({datasets: {dataset1: "TEST"}} as any)
+        });
+        await actions.changeLanguage({commit, dispatch, rootState} as any, Language.fr);
+
+        expectChangeLanguageMutations(commit);
+
+        expect(dispatch.mock.calls.length).toBe(2);
+        expect(dispatch.mock.calls[0][0]).toStrictEqual("metadata/getPlottingMetadata");
+        expect(dispatch.mock.calls[1][0]).toStrictEqual("genericChart/refreshDatasets");
     });
 
     it("changeLanguage fetches nothing if no relevant metadata to fetch", async () => {