+++ NEWS.md
@@ -1,3 +1,7 @@
+# hint 1.65.0
+
+* Show Input Time Series plots
+
 # hint 1.64.3
 
 * Fix language bug on reload
@@ -58,15 +62,29 @@
 
 * Save warnings from option validation, model fit and model calibration to app state
 
-# hint 1.56.1
+# hint 1.59.1
 
 * Render filter labels verbatim in output tables, do not attempt to translate
 
-# hint 1.56.0
+# hint 1.59.0
 
 * Include Back/Continue links at bottom of model options page
 
+# hint 1.58.0
+
+* Refresh output metadata and options when language changes
+
+# hint 1.57.0
+
+* Add tooltips for indicators in Time series chart
+
+# hint 1.56.0
+
+* Remove genericChart dataset when corresponding input file changes
+
 # hint 1.55.0
+
+* Add input Time series chart
 
 * Refresh output metadata and options when language changes